check unique flag on brightid verification return payload
