+++ vue-app/src/api/user.ts
@@ -4,15 +4,9 @@ import { Web3Provider } from '@ethersproject/providers'
 
 import { UserRegistry, ERC20 } from './abi'
 import { factory, ipfsGatewayUrl, provider } from './core'
-import { BrightIdError, getVerification } from './bright-id'
+import { BrightId } from './bright-id'
 
 export const LOGIN_MESSAGE = `Sign this message to access clr.fund at ${factory.address.toLowerCase()}.`
-
-export interface BrightId {
-  isLinked: boolean
-  isSponsored: boolean
-  isVerified: boolean // If is verified in BrightID
-}
 export interface User {
   walletAddress: string
   walletProvider: Web3Provider
@@ -59,37 +53,4 @@ export async function getEtherBalance(
   walletAddress: string
 ): Promise<BigNumber> {
   return await provider.getBalance(walletAddress)
-}
-
-export async function getBrightId(contextId: string): Promise<BrightId> {
-  const brightId: BrightId = {
-    isLinked: false,
-    isSponsored: false,
-    isVerified: false,
-  }
-
-  try {
-    await getVerification(contextId)
-    brightId.isLinked = true
-    brightId.isSponsored = true
-    brightId.isVerified = true
-  } catch (error) {
-    if (!(error instanceof BrightIdError)) {
-      /* eslint-disable-next-line no-console */
-      console.error(error)
-    }
-
-    // Not verified user
-    if (error.code === 3) {
-      brightId.isLinked = true
-      brightId.isSponsored = true
-    }
-
-    // Not sponsored user
-    if (error.code === 4) {
-      brightId.isLinked = true
-    }
-  }
-
-  return brightId
 }
